Add Accordion CSS demo (#503)
* Add CSS demo

* Convert in page examples to CSS

* Simplify and align demos with standard

* Tidy demo code

* Import order
