+++ components/demos/index.jsx
@@ -1,5 +1,5 @@
 export { GettingStartedDemo } from './GettingStarted';
-export { default as AccordionDemo } from './Accordion';
+export { default as AccordionDemo } from './Accordion/stitches';
 export { default as AlertDialogDemo } from './AlertDialog';
 export { default as AspectRatioDemo } from './AspectRatio';
 export { default as AvatarDemo } from './Avatar';
